refactor: created profileService
