+++ tests/unit/services/userService.test.ts
@@ -4,7 +4,6 @@ import { AccountStatus, AccountType, User } from '../../../src/entities/User';
 import users from '../../fixtures/users';
 import * as passwordUtils from '../../../src/util/password';
 import { getConnection, getRepository } from 'typeorm';
-import { APIProfile, Year, Course } from '../../../src/entities/Profile';
 import { HttpCode } from '../../../src/util/errors';
 
 beforeAll(async () => {
@@ -185,59 +184,6 @@ describe('UserService', () => {
 			await expect(userService.reportUser(reportingUser.id, reportedUser.id, {} as any)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
 			await expect(userService.reportUser(reportingUser.id, reportedUser.id, { description: null } as any)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
 			await expect(userService.reportUser(reportingUser.id, reportedUser.id, { description: 'hi' } as any)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
-		});
-	});
-	describe('putUserProfile', () => {
-		const userWithoutProfile = users.find(user => !user.profile)!;
-		const userWithProfile = users.find(user => user.profile)!;
-		beforeEach(async () => {
-			await getRepository(User).save([userWithProfile, userWithoutProfile]);
-		});
-
-		test('User profile registered for valid request (new profile)', async () => {
-			const savedUser = await userService.putUserProfile(userWithoutProfile.id, {
-				course: Course.COMPUTER_SCIENCE,
-				yearOfStudy: Year.TWO,
-				avatar: false,
-				instagram: '',
-				facebook: '',
-				twitter: '',
-				linkedin: ''
-			});
-			expect({ ...savedUser, profile: undefined }).toMatchObject(userWithoutProfile.toJSONPrivate());
-			const nonNullishProperties = [...Object.keys(savedUser.profile!)].filter(prop => savedUser.profile![prop as keyof APIProfile]);
-			expect(nonNullishProperties).toEqual(expect.arrayContaining(['id', 'course', 'yearOfStudy']));
-			expect(savedUser.profile!.course).toStrictEqual(Course.COMPUTER_SCIENCE);
-			expect(savedUser.profile!.yearOfStudy).toStrictEqual(Year.TWO);
-		});
-
-		test('User profile registered for valid request (existing profile)', async () => {
-			const initialProfile = userWithProfile.profile;
-			const savedUser = await userService.putUserProfile(userWithProfile.id, {
-				course: Course.SOFTWARE_ENGINEERING,
-				yearOfStudy: Year.ONE,
-				avatar: false
-			});
-			expect(userWithProfile).toMatchObject({ ...savedUser, profile: userWithProfile.profile });
-			const nonNullishProperties = [...Object.keys(savedUser.profile!)].filter(prop => savedUser.profile![prop as keyof APIProfile]);
-			expect(nonNullishProperties).toEqual(expect.arrayContaining(['id', 'course', 'yearOfStudy']));
-			expect(initialProfile).not.toMatchObject(savedUser.profile!);
-			expect(savedUser.profile!.course).toStrictEqual(Course.SOFTWARE_ENGINEERING);
-			expect(savedUser.profile!.yearOfStudy).toStrictEqual(Year.ONE);
-		});
-
-		test('Fails to create user profile for non-existent user', async () => {
-			const details = { course: Course.COMPUTER_SCIENCE, yearOfStudy: Year.TWO, avatar: false };
-			await expect(userService.putUserProfile('', details)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
-			await expect(userService.putUserProfile(`${userWithProfile.id}1`, details)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
-			await expect(userService.putUserProfile(`${userWithoutProfile.id}a`, details)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
-		});
-
-		test('Fails to create user profile with invalid details', async () => {
-			await expect(userService.putUserProfile(userWithoutProfile.id, {} as any)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
-			await expect(userService.putUserProfile(userWithoutProfile.id, { course: 'History', yearOfStudy: Year.ONE } as any)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
-			await expect(userService.putUserProfile(userWithoutProfile.id, { course: 'Software Engineering', yearOfStudy: 'Tenth Year' } as any)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
-			await expect(userService.putUserProfile(userWithoutProfile.id, { course: 'Computer Science', yearOfStudy: 3 } as any)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
 		});
 	});
 });
